requirements.txt
# AWS
boto3>=1.35.0
botocore>=1.35.0

# Anthropic direct (fallback mode)
anthropic>=0.40.0

# Data & validation
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI & logging
rich>=13.0.0
typer>=0.12.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
moto>=5.0.0  # AWS mocking for tests

# Dev
black>=24.0.0
ruff>=0.5.0
mypy>=1.10.0
